coin-analysis-app/requirements.txt
transformers
torch
pillow
requests
numpy
matplotlib
streamlit
opencv-python
